trying to fix build 1

+++ modules/task_1/Kryukov_num_of_altern_char/main.cpp
@@ -23,10 +23,10 @@ TEST(Parallel_Operations_MPI, Test_create_vector_with_correct_size) {
     int rank;
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
     std::vector<int> general_vec;
-    const long unsigned int size_vector = 10;
+    const int size_vector = 10;
     general_vec = getRandomVector(size_vector);
     if (rank == 0) {
-        ASSERT_EQ(general_vec.size(), 10);
+        ASSERT_EQ((int)general_vec.size(), 10);
     }
 }
